add baseUrlIssueBanner configuration
