+++ packages/docusaurus/src/client/baseUrlIssueBanner/styles.module.css
@@ -1,0 +1,3 @@
+.baseUrlIssueBanner {
+  display: none;
+}
