--- packages/docusaurus/src/client/baseUrlSuggestionWarning/styles.module.css
@@ -1,3 +1,0 @@
-.baseUrlSuggestionWarning {
-  display: none;
-}
